Upgraded alloc_name_srvc_pckt() to allocate an arbitrary number of fields.

+++ include/name_srvc_pckt.h
@@ -158,10 +158,10 @@ void *
   master_name_srvc_pckt_writer(struct name_srvc_packet *packet,
                                unsigned int *pckt_len);
 struct name_srvc_packet *
-  alloc_name_srvc_pckt(uint16_t qstn,
-                       uint16_t answ,
-                       uint16_t auth,
-                       uint16_t adit);
+  alloc_name_srvc_pckt(unsigned int qstn,
+                       unsigned int answ,
+                       unsigned int auth,
+                       unsigned int adit);
 void
   destroy_name_srvc_pckt(struct name_srvc_packet *packet,
                          unsigned int complete,
